Move json fix functions to a separate class
